@@ -1,14 +1,15 @@
 import { Errors, isExists, isString } from '@ridi/parser-core';
-import chai, { assert, should } from 'chai';
+import chai, { assert, expect, should } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import fs from 'fs-extra';
 import path from 'path';
+import sinon from 'sinon';
 
-import Book from '../src/model/Book';
+import ComicBook from '../src/model/ComicBook';
 import ComicParser from '../src/ComicParser';
-import Item from '../src/model/Item';
-import ReadContext from '../src/model/ReadContext';
-import ParseContext from '../src/model/ParseContext';
+import ComicItem from '../src/model/ComicItem';
+import ComicReadContext from '../src/model/ComicReadContext';
+import ComicParseContext from '../src/model/ComicParseContext';
 import Paths from '../../../test/paths';
 import validationBook from './validationBook';
 
@@ -21,10 +22,10 @@ describe('ComicParser', () => {
       const expectedContext = JSON.parse(fs.readFileSync(Paths.EXPECTED_COMIC_BOOK));
       let parser = new ComicParser(Paths.COMIC);
       let _context;
-    
+
       it('_prepareParse test', () => {
         return parser._prepareParse().then((context) => {
-          context.should.be.an.instanceOf(ParseContext);
+          context.should.be.an.instanceOf(ComicParseContext);
           context.options.should.deep.equal(ComicParser.parseDefaultOptions);
           context.entries.should.not.null;
           _context = context;
@@ -44,7 +45,7 @@ describe('ComicParser', () => {
           items.forEach((item) => {
             item.index.should.not.null;
             item.path.should.not.null;
-            item.fileSize.should.not.null;
+            item.size.should.not.null;
             isExists(item.width).should.be.false;
             isExists(item.height).should.be.false;
           });
@@ -58,13 +59,13 @@ describe('ComicParser', () => {
         });
       });
     });
-  
+
     it('Parse with default options from file', () => {
       return new ComicParser(Paths.COMIC).parse().then(book => {
         validationBook(book, JSON.parse(fs.readFileSync(Paths.EXPECTED_COMIC_BOOK)));
       });
     });
-  
+
     it('Parse with default options from directory', () => {
       return new ComicParser(Paths.UNZIPPED_COMIC).parse().then(book => {
         validationBook(book, JSON.parse(fs.readFileSync(Paths.EXPECTED_COMIC_BOOK)));
@@ -93,17 +94,29 @@ describe('ComicParser', () => {
       });
     });
   });
-  
+
+  it('initialize without cryptoProvider', () => {
+    const comicParser = new ComicParser(Paths.COMIC, 'fakeProvider');
+    expect(comicParser.cryptoProvider).to.be.undefined;
+  })
+
+  it('parseImageSize return undefined when sizeof throws', async () => {
+    const comicParser = new ComicParser(Paths.COMIC, 'fakeProvider');
+    const imageSize = await comicParser._parseImageSize({getFile:sinon.fake.returns(Buffer.from(['f','a','k','e']))}, {parseImageSize: true});
+    expect(imageSize.height).to.be.undefined;
+    expect(imageSize.width).to.be.undefined;
+  });
+
   describe('Book serialization test', () => {
     it('Book -> RawBook -> Book', () => {
       return new ComicParser(Paths.COMIC).parse().then(book => {
         const rawBook = book.toRaw();
-        const newBook = new Book(rawBook);
+        const newBook = new ComicBook(rawBook);
         validationBook(book, JSON.parse(fs.readFileSync(Paths.EXPECTED_COMIC_BOOK)));
       });
     });
   });
-  
+
   describe('Reading test', () => {
     const parser = new ComicParser(Paths.COMIC);
     parser.parse().then(book => {
@@ -113,7 +126,7 @@ describe('ComicParser', () => {
             assert(Buffer.isBuffer(result));
           });
         });
-  
+
         it('Read single item with base64', () => {
           return parser.readItem(book.items[0], { base64: true }).then((result) => {
             const expected = fs.readFileSync(Paths.COMIC_BASE64, 'utf-8');
@@ -121,7 +134,7 @@ describe('ComicParser', () => {
             result.substr(0, 500).should.equal(expected.substr(0, 500));
           });
         });
-  
+
         it('Read single item with force', () => {
           return parser.readItem({ path: '1.jpg' }, { force: true }).then((result) => {
             const expected = fs.readFileSync(Paths.COMIC_FIRST);
@@ -129,7 +142,7 @@ describe('ComicParser', () => {
             result.equals(expected).should.be.true;
           });
         });
-  
+
         it('Read multiple items', function () {
           this.timeout(10 * 1000);
           return parser.readItems(book.items).then((results) => {
