@@ -1,7 +1,7 @@
 import { assert, should } from 'chai';
 
 import CssItem from '../../src/model/CssItem';
-import Item from '../../src/model/Item';
+import BaseEpubItem from '../../src/model/BaseEpubItem';
 
 should(); // Initialize should
 
@@ -9,7 +9,7 @@ describe('Model - CssItem', () => {
   it('Initialize test', () => {
     let item = new CssItem();
     assert(item.namespace === undefined);
-    assert(item instanceof Item);
+    assert(item instanceof BaseEpubItem);
 
     item = new CssItem({ id: 'style.css', href: './style.css', mediaType: 'text/css', size: 15, namespace: 'css_1' });
     item.namespace.should.equal('css_1');
