@@ -1,6 +1,6 @@
 import { assert, should } from 'chai';
 
-import Item from '../../src/model/Item';
+import BaseEpubItem from '../../src/model/BaseEpubItem';
 import NcxItem from '../../src/model/NcxItem';
 
 should(); // Initialize should
@@ -9,7 +9,7 @@ describe('Model - NcxItem', () => {
   it('Initialize test', () => {
     let item = new NcxItem();
     item.navPoints.should.have.lengthOf(0);
-    assert(item instanceof Item);
+    assert(item instanceof BaseEpubItem);
 
     item = new NcxItem({ navPoints: [{ id: 'c1' }, { id: 'c2' }] });
     item.navPoints.should.have.lengthOf(2);
