@@ -1,5 +1,10 @@
-import es from 'event-stream';
+import * as es from 'event-stream';
 
+/**
+ * @param {number} start=0
+ * @param {number} end=Infinity
+ * @returns {es.MapStream}
+ */
 const createSliceStream = (start = 0, end = Infinity) => {
   let bytesReceived = 0;
   let finish = false;
