@@ -1,5 +1,11 @@
 import { isExists, isObject } from './typecheck';
-
+/**
+ * Merge Objects
+ * @param {T} obj1
+ * @param {S} obj2
+ * @returns {K} merged object
+ * @template T,S,K
+ */
 export default function mergeObjects(obj1, obj2) {
   return [obj1, obj2].reduce((merged, obj) => {
     Object.keys(obj).forEach((key) => {
