@@ -1,6 +1,10 @@
 import { isBool, isString } from './typecheck';
 import { stringContains } from './stringUtil';
-
+/**
+ * Return boolean form of any input
+ * @param {any} any
+ * @returns {boolean}
+ */
 export default function parseBool(any) {
   if (isBool(any)) {
     return any;
