@@ -1,6 +1,4 @@
-/* eslint-disable max-len */
 import fs from 'fs-extra';
-import StreamChopper from 'stream-chopper';
 import AdmZip from 'adm-zip';
 import { Readable } from 'stream';
 
@@ -12,31 +10,52 @@ import { isExists } from './typecheck';
 import { conditionally } from './streamUtil';
 import Errors, { createError } from './errors';
 
+/**
+ * @typedef GetFileOptions
+ * @property {string} encoding
+ * @property {number} end
+ */
+
+/**
+ * @typedef {Object} ZipFileInformation
+ * @property {string} file
+ * @property {AdmZip.IZipEntry[]} files
+ * @property {CryptoProvider} cryptoProvider
+ * @property {(entryPath: string) => AdmZip.IZipEntry} find
+ * @property {(entry: AdmZip.IZipEntry, options?: GetFileOptions) => Promise<Buffer | String>} getFile
+ * @property {(unzipPath: string, overwrite?: boolean) => Promise<any>} extractAll
+ * @property {import('./Logger').default} logger
+ */
+
+/**
+ * Find the file with a path.
+ * @param {string} entryPath File Path
+ * @returns {AdmZip.IZipEntry | undefined} A file with path or undefined if there is none.
+ * */
 function find(entryPath) {
   return this.files.find(entry => entryPath === entry.entryName);
 }
 
-function _getBufferSize(cryptoProvider) {
-  if (isExists(cryptoProvider)) {
-    return cryptoProvider.bufferSize;
-  }
-  return undefined;
-}
-
-async function getFile(entry, options = {}) {
+/**
+ *
+ * @async
+ * @this {ZipFileInformation}
+ * @param {import('adm-zip').IZipEntry} entry
+ * @param {GetFileOptions} options
+ * @returns {Promise<Buffer | String>} String is encoding is provided, Buffer otherwise
+ */
+async function getFile(entry, options = { encoding: undefined, end: undefined }) {
   const { encoding, end } = options;
-  let file = await new Promise((resolve, reject) => {
+  let file = await new Promise((resolveFile) => {
     const totalSize = Math.min(end || Infinity, entry.uncompressedSize);
-    const bufferSize = _getBufferSize(this.cryptoProvider);
     let data = Buffer.from([]);
     const readable = Readable.from(entry.getData());
     readable
-      .pipe(conditionally(isExists(bufferSize), new StreamChopper({ size: Math.min(bufferSize, entry.uncompressedSize) })))
       .pipe(conditionally(isExists(end), createSliceStream(0, end)))
-      .pipe(conditionally(this.cryptoProvider && !!this.cryptoProvider.isStreamMode, createCryptoStream(entry.path, totalSize, this.cryptoProvider, CryptoProvider.Purpose.READ_IN_ZIP)))
+      .pipe(conditionally(this.cryptoProvider && !!this.cryptoProvider.isStreamMode,
+        createCryptoStream(entry.path, totalSize, this.cryptoProvider, CryptoProvider.Purpose.READ_IN_DIR)))
       .on('data', (chunk) => { data = Buffer.concat([data, chunk]); })
-      .on('error', e => reject(e))
-      .on('end', () => resolve(data));
+      .on('end', () => { resolveFile(data); });
   });
   if (this.cryptoProvider && !this.cryptoProvider.isStreamMode) {
     file = this.cryptoProvider.run(file, entry.path, CryptoProvider.Purpose.READ_IN_DIR);
@@ -50,6 +69,13 @@ async function getFile(entry, options = {}) {
   return file;
 }
 
+/**
+ * Extract zip file to path
+ * @this {ZipFileInformation}
+ * @param {string} unzipPath Path where files will be extracted
+ * @param {boolean} overwrite
+ * @returns {Promise<void>}
+ */
 async function extractAll(unzipPath, overwrite = true) {
   if (overwrite) {
     fs.removeSync(unzipPath);
@@ -65,6 +91,13 @@ async function extractAll(unzipPath, overwrite = true) {
   zip.extractAllTo(unzipPath);
 }
 
+/**
+ * @param {string | Buffer} file
+ * @param {CryptoProvider} cryptoProvider
+ * @param {Logger} logger
+ * @returns {ZipFileInformation}
+ * @throws {Errors.ENOENT} When file can't be found
+ */
 export default async function openZip(file, cryptoProvider, logger) {
   try {
     const files = new AdmZip(file).getEntries();
