@@ -1,9 +1,13 @@
-import { should } from 'chai';
+import { expect, should } from 'chai';
 import fs from 'fs-extra';
 
 import Errors from '../src/errors';
 import openZip from '../src/zipUtil';
 import Paths from '../../../test/paths';
+import TestSyncCryptoProvider from './testSyncCryptoProvider';
+import TestSyncStreamCryptoProvider from './testSyncStreamCryptoProvider';
+import TestAsyncCryptoProvider from './testAsyncCryptoProvider';
+
 
 should(); // Initialize should
 
@@ -26,6 +30,10 @@ describe('Util - Zip', () => {
     const zip = await openZip(Paths.DEFAULT);
     await zip.extractAll('./temp');
   });
+  it('extractAll test with cryptoprovider', async () => {
+    const zip = await openZip(Paths.DEFAULT, new TestSyncCryptoProvider);
+    await zip.extractAll('./temp');
+  });
 
   it('extractAll with not overwite', async () => {
     const zip = await openZip(Paths.DEFAULT);
@@ -42,4 +50,24 @@ describe('Util - Zip', () => {
       });
     });
   });
+  describe('works with CryptoProvider', () => {
+    it(': AsyncCryptoProvider', async () => {
+      const testCryptoProvider = new TestAsyncCryptoProvider();
+      const zip = await openZip(Paths.DEFAULT, testCryptoProvider);
+      const entry = zip.find('mimetype');
+      entry.should.be.not.undefined.and.not.null;
+      const buffer = await zip.getFile(entry);
+      Buffer.isBuffer(buffer).should.be.true;
+      expect(buffer).to.deep.equal(entry.getData());
+    })
+    it(': SyncCryptoProvider', async () => {
+      const testCryptoProvider = new TestSyncCryptoProvider();
+      const zip = await openZip(Paths.DEFAULT, testCryptoProvider);
+      const entry = zip.find('mimetype');
+      entry.should.be.not.undefined.and.not.null;
+      const buffer = await zip.getFile(entry);
+      Buffer.isBuffer(buffer).should.be.true;
+      expect(buffer).to.deep.equal(entry.getData());
+    })
+  })
 });
